requirements.txt
requests==2.23.0
python-telegram-bot==12.7
python-dotenv==0.13.0
PySocks==1.7.1
